Write the coordinator records Kafka's decoder can read
The `__consumer_offsets` evidence suite handed krabka's topic to Kafka 4.3.1's
own formatter and the decoder died on the first KIP-848 record:
`BufferUnderflowException` inside `ConsumerGroupMetadataValue.read`. Every
value of the KIP-848, KIP-932 and KIP-1071 families is `flexibleVersions: 0+`
in Kafka's schema -- compact strings, compact arrays, and a tagged-field
trailer on the message and each nested struct -- and krabka wrote all of them
in the old non-flexible shape. The keys are not flexible (`"none"` at 4.3.1),
so they are unchanged.

The key-version discriminators were wrong too. krabka numbered the share and
streams families from where its own previous record left off, so it wrote key
16 meaning streams member metadata where Kafka reads
`ConsumerGroupRegularExpression`, and 17 meaning streams topology where Kafka
reads `StreamsGroupMetadata`. Those are types Kafka's serde knows, so its
tools do not skip them: they mis-decode or die mid-topic. Every record now
carries Kafka's own `apiKey`; 4, 9 and 16 are left unused, and the streams
partition-metadata snapshot stays on 18, which KIP-1101 retired and Kafka
therefore skips as unknown.

Reading the schemas end to end turned up more divergences than the underflow:
`MemberState` discriminants were swapped, streams member state counts from 1
in Kafka and from 0 here, `ConsumerGroupMemberMetadataValue` and
`StreamsGroupMemberMetadataValue` had fields in the wrong order, the streams
endpoint port is a `uint16`, topic ids were length-prefixed where the wire
carries 16 raw bytes, subtopology changelog topics precede the repartition
sinks, and several fields Kafka always writes were absent. All are corrected,
and the fields the broker keeps no state for are written with Kafka's own
defaults and dropped on decode, each documented at its record.

coordinator/unified/persistence/flex.rs is the one shared flexible codec,
bound to `krabka_protocol`'s primitives and tagged-field readers rather than
hand-rolled. Every record has a test comparing whole expected bytes against
Kafka's schema, a round trip, and a case that fails if the tagged trailer is
ever dropped again -- the byte-exact stand-in for a container suite this
environment cannot run.

+++ crates/broker/src/coordinator/unified/streams/state/member.rs
@@ -14,9 +14,12 @@ use krabka_log::Offset;
 /// It mirrors KIP-848's `MemberAssignmentState`. Standby and warmup tasks take
 /// no part in it.
 ///
-/// Persistence stores this as a raw `i8`. [`as_i8`](Self::as_i8) and
-/// [`from_i8`](Self::from_i8) convert it without coupling this module to the
-/// persistence layer.
+/// This is the broker's own numbering, which counts from zero. Kafka's
+/// `org.apache.kafka.coordinator.group.streams.MemberState` counts from one, so
+/// the persistence layer converts through its own `StreamsMemberWireState`
+/// rather than storing these discriminants. [`as_i8`](Self::as_i8) and
+/// [`from_i8`](Self::from_i8) expose the discriminant for the model checks that
+/// compare states as numbers.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub enum StreamsMemberAssignmentState {
     /// The member's active tasks match its target, and nothing is pending.
@@ -71,7 +74,7 @@ pub struct StreamsMemberState {
     pub process_id: String,
     /// Optional `(host, port)` the member advertises for interactive-query
     /// routing.
-    pub user_endpoint: Option<(String, u32)>,
+    pub user_endpoint: Option<(String, u16)>,
     /// Arbitrary `(key, value)` client tags for rack-aware and custom
     /// assignment.
     pub client_tags: Vec<(String, String)>,
